Storage: Fix moving apps to external storage
This commits reverts the changes to PackageHelper.java made by commit
99d3b9 "Storage Add secondary storage support for app installation" to
fix moving/installing apps on the external storage for devices where the
secondary storage is used for USB OTG and not always mounted.

In commit 99d3b9 the logic was changed to check the secondary storage
without a fallback to the external storage in case of failure. Furthermore
the whole change seems incomplete as other parts of the package manager only
use the external storage.

Change-Id: I017899114fca2fe3c789e07dc74f8fa165df61d1